==> 02-data.Rmd <==
# Data sources

## Colonial History Dataset

**Source :**
Paul R. Hensel (2018). "ICOW Colonial History Data Set, version 1.1." Available at <http://www.paulhensel.org/icowcol.html>.

The colonial history data set contains colonial information for over 200 countries, including, if they were colonized, their principal Colonial Ruler, the years that this colonial domination began and formally ended. Several colonies were controlled by different colonial powers over the years, in our analysis we focused on the principal colonial ruler, as designated by the dataset, on the assumption that that country would have had the biggest effect on the colony.

**Variables of the Colonial History Data Set**

```{r echo=FALSE}
coldata <- read.csv("Data/raw/coldata110.csv")
colnames(coldata)
```

## Current Indicators Data (GDP, Gini, HDI)

Worldwide GDP per capita data in 2017. 
Source: [databank.worldbank.org](https://databank.worldbank.org/home.aspx). 

We obtained data on Gini coefficient and the human development index from the UNDP, you can find the information here at [hdr.undp.org](http://hdr.undp.org/en/data)
The actual download we used is a little difficult to find so the direct link is [here](https://www.google.com/url?sa=t&rct=j&q=&esrc=s&source=web&cd=3&ved=2ahUKEwj92b7u7LDmAhWWVs0KHY5-Dy4QFjACegQIBRAC&url=http%3A%2F%2Fhdr.undp.org%2Fsites%2Fdefault%2Ffiles%2F2018_all_indicators.xlsx&usg=AOvVaw1X2alKbj6XdNRcUGJ_DD-E)
Please note that this is a direct link to the download so only click on it if you would like to download a copy of the data.

We pulled the most recent information we could find for each of these countries in an attempt to think about how colonization may have impacted what is going on in a country today. Of course the time period of colonization was extremely long and complicated, and about 70 years have passed since colonialism ended, so many other factors certainly have an impact. But this was the concept we were hoping to explore.

per capita GDP 2017: Represents gross domestic product per capita in 2017.
Gini Coefficient: Measurement of inequality, the higher the value, the more unequal the society. These measurements were taken between 2010-2017. There was only one measurement per country, but which year that measurement came from was unclear. We felt that given the span of colonialism, this was acceptable as a method to evaluate a country's current status.

HDI: Stands for human development index. A higher value represents a higher quality of life. Details regarding how this index is calculated and what it measures can be found [here](http://hdr.undp.org/en/content/human-development-index-hdi)

##  Angus Maddison Dataset: GDP Over Time

The dataset we used to look at GDP over time can be found [here](https://www.rug.nl/ggdc/historicaldevelopment/maddison/)

This dataset consists of countries and their estimated per capita GDP's overtime. The Maddison dataset comes from a project has been developing Maddison's work since his death. These numbers are the most updated versions including recent historical work. The links to the sources provide detailed documentation/explanation of the changes and significance. They provide two types of per capita GDP, one consistent with growthrates and another that ensured the comparative value of goods in a single year was consistent. We chose to use the latter measurement since it seemed more appropriate given our interest in comparing per capita GDP, rather than growth rate. This decision was made based on the documentation provided with the data.

We used this data to see if we could look at the economic impact of colonialism overtime and see any patterns. Ideally we would have liked to look at the per capita GDP for a few countries prior to colonization, during colonization, and after colonization. The prior to colonization time period and some of the years during colonization, proved to be difficult as for the vast majority of countries this information is unavailable. In fact, yearly data for many countries only begins around 1950, when colonization ended. This makes sense as the world looked very different then, modern countries frequently didn't exist, and the level of economic integration. However, it did limit our ability to analyze changes over time. We did our best to use what we had to investigate how countries' economic situation changed over time and possible relationships with colonialism.

## Colonial Transformation Dataset

**Source :**
[Research from Ziltener, Patrick & Künzler, Daniel & Walter, André. (2017). Measuring the Impacts of Colonialism: A New Data Set for the Countries of Africa and Asia. Journal of World-Systems Research. 23. 156-190. 10.5195/jwsr.2017.683.](https://www.researchgate.net/publication/314156755_Measuring_the_Impacts_of_Colonialism_A_New_Data_Set_for_the_Countries_of_Africa_and_Asia) 

This Dataset  encompasses 83 Asian and African countries and ranks them according to 15+ indicators in an effort to quantify the type of colonial domination they experienced. The indicators are divided between Political (with a sub-category on violence), Economic, and Social. This data is of a numerical/categorical hybrid nature, as they are numerical encoding of non-numerical (but gradual) realities. (For instance, "Colonial Rule" has been encoded on a scale of 0-5 from 0 for no colonial rule to 5 for direct rule and complete interference from foreign countries.) A summary of all those variables can be found in the github files under src.

**Variables of the Colonial Transformations Data Set**
```{r echo=FALSE}
transfo_raw <- read.csv("Data/raw/Colonial_transformation_data.csv")

colnames(transfo_raw)
```
