==> 07-conclusion.Rmd <==
# Conclusion

As we explored the data, we felt narrowing the scope at the beginning of the project would have allowed us to investigate a specific relationship more closely, as there seemed to be no shortage of historical perspectives to look at. We found ourselves somewhat overwhelmed as to where to look more closely. But, narrowing it down also required going deeper in domain knowledge/specific history and that seemed to be outside the scope of this project.

What we found was there do seem to be some patterns between colonialism and per capita GDP. This came accross most clearly when looking comparing average per capita GDP by colonizing country as well as how long a country was colonized and current per capita GDP. Of course the number of events that occurred during colonialism and time afterward have been so large and often incredibly significant changes that making any kind of strong statement on the impact of colonialism on modern day wasn't possible. We did find it interesting to note that colonial countries with high per capita GDPs today have very well known reasons for those circumstances. But our attempts to suss out the impact of colonialization by examining GDP per capita over time was somewhat stymied by the lack of data for the crucial time period of colonization. However, we did note a few interesting trends that point to areas we could investigate to understand the history of colonialism better.

We struggled at times with not finding clearer trends, but one of the most important take-aways from this class was how easy it is to stretch and squeeze the data to make trends appear, and if we feel like we have to do so, then we might be looking to confirm our pre-existing assumptions instead of genuinely exploring the data. 

We were also confronted with the limitations of our data itself. We worked almost exclusively with man-made data, rather than empirical measurements. From the GDP to the indicators, all of our parameters, formula or assessment-based, included human decision and subsequently, bias. In addition, we wanted to look at how the data changed overtime, data which we often did not have, had in an unequal fashion for different observations, or the accuracy of which was very uncertain (the GDP data prior to certain year, for instance, was very approximate). 

With all of those caveats, the colonial datasets remain rich ressources that try to make sense, in a quantifiable way, of a very complex and time-spanning issue. We hope that our vizualizations contribute to that goal. While  our results are inconclusive, we have learned a lot from working with this data, getting a better sense of the factors historians might look at more extensively, and the issues they may face regarding quantitative support of their investigations. We have a better sense of the how quantitative analysis could contribute to historical/humanities work factors, as well as its limitations.
